# Special case strings in System.Linq.Enumerable.OrderBy