--- a/src/System.Linq/src/System/Linq/OrderedEnumerable.cs
+++ b/src/System.Linq/src/System/Linq/OrderedEnumerable.cs
@@ -172,11 +172,11 @@ namespace System.Linq
             // Special case the common use of string with default comparer. Comparer<string>.Default checks the
             // thread's Culture on each call which is an overhead which is not required, because we are about to
             // do a sort which remains on the current thread (and EnumerableSorter is not used afterwards).
-            var comparer =
-                typeof(TKey) == typeof(string) // should always be true if => _comparer == Comparer<string>.Default...
-                && _comparer == Comparer<string>.Default
-                ? (IComparer<TKey>)StringComparer.CurrentCulture
-                : _comparer;
+            IComparer<TKey> comparer = _comparer;
+            if (typeof(TKey) == typeof(string) && comparer == Comparer<string>.Default)
+            {
+                comparer = (IComparer<TKey>)StringComparer.CurrentCulture;
+            }
 
             EnumerableSorter<TElement> sorter = new EnumerableSorter<TElement, TKey>(_keySelector, comparer, _descending, next);
             if (_parent != null)