First in possibly a series of performance related pull requests to enhance performance of System.Linq. (Depending on how this is received...)

There changed originated from [Cistern.Linq](https://github.com/manofstick/Cistern.Linq) which is a complete overhaul of System.Linq architecture, but is probably a [bit too much of a change](https://github.com/dotnet/corefx/pull/34208) for corefx to bear!

So anyway, this PR special cases string handling in OrderBy (when the default comparer is used - i.e. no extra parameter). I think this is a pretty common situation (although your telemetry might inform otherwise?). It increases performance because the comparer doesn't need to constantly get the current threads culture - which is consistent for the period of the sort, so an unnecessary overhead.

For strings like that of a person's name on a person object, this cuts about ~20% off the runtime, for a  slow-down when there is only a single item to be sorted - I assume due to my additional check and replacement of the comparer, but could of just been an artifact of the particular run? Would be good for some other verification of alternative machines... Anyone?

Following [benchmarking procedures](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md), the following results were obtained:

(The additional performance test that is used here is [in a PR here](https://github.com/dotnet/performance/pull/992))

| Slower                                                             | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.Linq.Tests.Perf_OrderBy.OrderByFirstName(NumberOfPeople: 1) |      1.40 |           187.96 |           263.96 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFullName(NumberOfPeople: 1)  |      1.22 |           352.27 |           430.88 |         |

| Faster                                                                | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Linq.Tests.Perf_OrderBy.OrderByFirstName(NumberOfPeople: 100)  |      1.26 |         52061.62 |         41456.28 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFirstName(NumberOfPeople: 1000) |      1.24 |        805309.06 |        647744.40 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFirstName(NumberOfPeople: 10)   |      1.23 |          2538.67 |          2060.65 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFullName(NumberOfPeople: 1000)  |      1.19 |       1020842.19 |        859375.00 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFullName(NumberOfPeople: 100)   |      1.18 |         65284.81 |         55141.30 |         |
| System.Linq.Tests.Perf_OrderBy.OrderByFullName(NumberOfPeople: 10)    |      1.07 |          3710.74 |          3481.37 |         |