@stephentoub 2019-11-01
I suspect we don't have any tests that validate culture behavior for string sorts / element ats / etc., including when an enumerable is enumerated multiple times with a different current culture each time, or when the culture changes between calls to MoveNext.  Can you please add some?