@manofstick 2019-11-02
@stephentoub 

Hmm. I *think* that there is an existing culture sorting issue that is unrelated to my PR on the Linux and Mac builds (that is I'm assuming that the Windows build does the correct thing, as the tests for me pass on windows on both core and framework - the tests were derived from the [Sorting and String Comparison](https://docs.microsoft.com/en-us/globalization/locale/sorting-and-string-comparison) online docs).

I have [conditionally excluded my change](https://github.com/dotnet/corefx/pull/42286/files#diff-d90239bd8284188e2bd210790483f5caR176-R181) but get [test failures](https://helix.dot.net/api/2019-06-17/jobs/ee544c28-c451-47f7-8359-57ef93da6874/workitems/System.Linq.Tests/console) such as this on the mac build:

```
    System.Linq.Tests.OrderByTests.CultureOrderByElementAt [FAIL]
      Assert.Equal() Failure
      Expected: Apple0
      Actual:   ï¿½ble0
      Stack Trace:
        /_/src/System.Linq/tests/OrderByTests.cs(523,0): at System.Linq.Tests.OrderByTests.CultureOrderByElementAt()
    System.Linq.Tests.OrderByTests.CultureOrderBy [FAIL]
      Assert.Equal() Failure
      Expected: String[] ["Apple0", "Apple1", "Apple2", "ï¿½ble0", "ï¿½ble1", ...]
      Actual:   String[] ["ï¿½ble0", "ï¿½ble1", "ï¿½ble2", "Apple0", "Apple1", ...]
      Stack Trace:
        /_/src/System.Linq/tests/OrderByTests.cs(469,0): at System.Linq.Tests.OrderByTests.CultureOrderBy()
```

The [windows build failure](https://helix.dot.net/api/2019-06-17/jobs/1839ec66-2f16-479e-b130-c3ec3630b8f2/workitems/System.Net.NetworkInformation.Functional.Tests/console) seems unrelated also, *especially* given that my change is excluded...