@stephentoub 2019-11-04
> I think that there is an existing culture sorting issue that is unrelated to my PR on the Linux and Mac builds (that is I'm assuming that the Windows build does the correct thing

.NET Core relies on the OS / underlying globalization support for culture-specific sorting tables.  Currently on Windows that means using the data from Windows, and on Unix it means using ICU, and their sorting data can differ from each other. cc: @tarekgh