@tarekgh 2019-11-04
> I think that there is an existing culture sorting issue that is unrelated to my PR on the Linux and Mac builds (that is I'm assuming that the Windows build does the correct thing

Something you may try and look if it will help, when calling OrderBy try to pass a comparer with ignore symbols. something like: 

```C#

StringComparer comparer = StringComparer.Create(new CultureInfo("da-DK"), CompareOptions.IgnoreSymbols);
source.OrderBy(x => x, comparer)

```