@manofstick 2019-11-05
@tarekgh

Thanks for having a look, but the point of this PR is to handle the case where no comparer is passed to `OrderBy` , so to optimize the path of `Comparer<string>.Default` which will use the `CurrentCulture`, but as the thread doesn't change during the whole sort it only needs to access it once.